Refactored display
